LayoutTests/accessibility/mac/link-with-title-expected.txt

TEXT
This tests that a link with a title attribute will only include that in the help field.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Link with title and visible text
PASS link1.description is 'AXDescription: '
PASS link1.title is 'AXTitle: TEXT'
PASS link1.helpText is 'AXHelp: TITLE'
PASS successfullyParsed is true

TEST COMPLETE
